Revert "change readme sample instructions in line with the hexo-theme-meteor repo"
This reverts commit 32b8c00e5a688a492b23fce21919c287642aca18.

Actually, @evans, can you rename your `hexo-theme-meteor` folder (which you
have locally) to `meteor-theme-hexo`?  I just made this change intentionally
a few hours ago in #11.

This was an intentional change to the name of the actual repository, so that
it matched the name of the real npm (which is `meteor-theme-hexo`) and thus
allows `npm link` to work properly.