Manually fix package-lock.json to include `is-wsl` as a dependency of `opn`.
The `is-wsl` package is not an optional dependency, nor it it confined by
any `engine` rule which this repository doesn't comply with.  In fact,
strangely, it just doesn't get installed as a dependency of `opn` when it is
in fact the _only_ dependency of that package and very much _required_ to
run.

There's certainly been a transitive dependency change somewhere which is
causing this to happen, but I honestly have to chalk it up to a bug in npm
since running an `npm ls` immediately after an `npm install` results in the
absence of `is-wsl` being immediately flagged as a problem.

Aims to fix https://github.com/meteor/meteor-theme-hexo/pull/86.